Unify session and interop token in the same file
